pyproject.toml
[project]
name = "sentinel-fraud-env"
version = "0.1.0"
description = "SENTINEL Fraud Investigation Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "openai>=1.40.0",
    "python-multipart",
    "huggingface_hub[cli]>=0.25.0",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
